requirements.txt
biopython==1.81
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
tzdata==2023.3
zipp==3.17.0
